Switch over to explicitly generating README.md from org file
The problem is that GitHub does display Org files, but not very well.
It is better to generate a markdown README file myself, so I have
control over how it comes out.  This is the only way I have found to
get the links from the README to the jupyter notebooks to work.