Revisit question of He ionization in the bow shock
Redo the calculations from a slightly different point of view. I seem
to be converging on some consistent answer, but there are still some
puzzling results that don't seem to fit. For instance calculating the
electron density from the H beta line gives too low a value, given
that we think the ionization fraction of He++ is low.

This still needs more work to resolve it and get everything
consistent.