Progress on the QA figure but hit snag with hyper-local continuum
It works fine for the spectra, but the images and moments seem to not
change at all when I turn on/off the subtraction of the hyper-local
continuum. I need to investigate this further. I plan to remove the
compare-two-maps.py script from the org file and edit it directly as a
python file